requirements.txt
# Core ML and NLP libraries
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# FAISS for vector search
faiss-cpu>=1.7.4

# LLM support
llama-cpp-python>=0.2.0
huggingface-hub>=0.16.0

# PDF processing
pdfminer.six>=20221105

# Text processing
nltk>=3.8.0

# Standard libraries and utilities
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.0.0
